fix need run calc

+++ khelo/lib/ui/flow/score_board/score_board_view_model.dart
@@ -1432,7 +1432,7 @@ class ScoreBoardViewNotifier extends StateNotifier<ScoreBoardViewState> {
     super.dispose();
   }
 
-  Future<void> setRevisedTarget(int run, int over) async {
+  Future<void> setRevisedTarget(int run, double over) async {
     String? matchId = state.match?.id;
     if (matchId == null) {
       return;
